+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# New
+
+* Add basic logging layer
+* Internal cleanup
+* Drop grape from gemspec
+
+# 1.2.2
+
+* Bind address was not actualy bound
+
+# 1.2.1
+
+* Fix pechkin error when auth header contains invalid content.
+* Add auth error messages
+* Remove debug oputput from Auth Middleware
+
 # 1.2.0
 
 * Add bind address support